api/requirements.txt
black==22.3.0
certifi==2022.6.15
click==8.1.3
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.22.4
pandas==1.4.2
pathspec==0.9.0
platformdirs==2.5.2
polygon-api-client==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
websockets==10.3
Werkzeug==2.1.2
zipp==3.8.0
